serving/requirements.txt
# Web
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3

# Data / ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.7.2
joblib==1.4.2
lightgbm==4.3.0

# AWS / S3
boto3==1.34.69
botocore==1.34.69

# Utils
python-dotenv==1.0.1
pydantic==2.6.4

#etc
requests